Bug regarding cart has been fixed.
Notes:
 - Improvements in calling the mobile page not found view. (not implemented in all functions yet).
 - Improvements in calling neccessary navbar data.

+++ database/migrations/2021_05_06_081917_create_orders_table.php
@@ -20,7 +20,7 @@ class CreateOrdersTable extends Migration
             $table->unsignedBigInteger('course_id')->nullable();
             $table->foreign('course_id')->references('id')->on('courses')->onDelete('set null');
             $table->integer('qty');
-            $table->integer('price');
+            $table->unsignedBigInteger('price');
             $table->boolean('withArtOrNo')->default(0)->nullable();
             $table->timestamps();
         });
